Extract RawResultRow to separate file
